refactor(cmd/mdsmith): move list-query subcommand logic into query.go
Move parseQueryFlags, runQuery, queryFiles, readFrontMatterRaw, and
queryOptions from main.go into a new cmd/mdsmith/query.go, matching the
pattern already used for backlinks.go and init.go. Move their unit tests
from main_unit_test.go into a new query_unit_test.go.

No behaviour change. Closes plan/2608021915.
